requirements.txt
alembic==1.13.1
anyio==4.3.0
APScheduler==3.10.4
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email-validator==2.1.0.post1
Flask==3.0.2
Flask-Admin==1.6.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
idna==3.6
infinity==1.5
intervals==0.9.2
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
polygon-api-client==1.13.4
PyPika==0.48.9
pytz==2024.1
requests==2.31.0
schedule==1.2.1
setuptools==69.1.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.27
SQLAlchemy-Utils==0.41.1
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
urllib3==1.26.18
validators==0.22.0
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.1
WTForms==3.1.2
WTForms-Alchemy==0.18.0
WTForms-Components==0.10.5
